+++ titanjr-dubbo-api/src/main/java/com/fangcang/titanjr/dto/bean/TitanTransferDTO.java
@@ -19,8 +19,8 @@ public class TitanTransferDTO implements Serializable {
     private String intermerchantcode;
     private String userrelateid;
     private String interproductid;
-    private Double amount;
-    private Double userfee;
+    private Long amount;
+    private Long userfee;
     private String requestno;
     private Date requesttime;
     private String useripaddress;
@@ -115,19 +115,19 @@ public class TitanTransferDTO implements Serializable {
         this.interproductid = interproductid;
     }
 
-    public Double getAmount() {
+    public Long getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(Long amount) {
         this.amount = amount;
     }
 
-    public Double getUserfee() {
+    public Long getUserfee() {
         return userfee;
     }
 
-    public void setUserfee(Double userfee) {
+    public void setUserfee(Long userfee) {
         this.userfee = userfee;
     }
